requirements.txt
tensorflow[and-cuda]
numpy
matplotlib
scikit-learn
pillow
seaborn
jupyter
